machines: clear the -Wall -Wextra warnings
Behaviour-neutral:
- sign-compare in av, civic, se30, mac030_glue, mac030_rom_overlay, mcu,
  iici, iisi, iifx: every one a leftover (int)page < g_page_count after
  g_page_count became uint32_t; the casts go.
- iix, se30, oss, checkpoint_images and tnt/swim3 declared a log
  category and never logged.
- iifx: scsidma_fifo_put_lane/get_lane and scsidma_lane_shift had no
  callers. (So nothing ever puts a byte into the IIfx SCSI-DMA FIFO, and
  the fifo_count checks around it cannot fire -- true before this change
  too; recorded, not changed.)

+++ src/machines/av/av.c
@@ -458,7 +458,7 @@ static void av_scc_irq(void *context, bool active) {
 static void av_map_ram(config_t *cfg) {
     uint8_t *ram_base = ram_native_pointer(cfg->mem_map, 0);
     uint32_t pages = cfg->ram_size >> PAGE_SHIFT;
-    for (uint32_t p = 0; p < pages && (int)p < g_page_count; p++)
+    for (uint32_t p = 0; p < pages && p < g_page_count; p++)
         mac030_fill_page(p, ram_base + (p << PAGE_SHIFT), true);
 }
